[B] Better handle page copying/moving
There are issues with the TYPO3 backend when copying and when moving
pages that have Short URIs on them:

When copying a page: the short URIs attached to the page should not get
copied along with it. This commit ensures that doesn't happen. Under the
hood the simplest way to do so is to simply delete the newly created
Short URI copy after everything's said and done in DataHandler.

When moving a page: somehow with IRRE relations, TYPO3 doesn't end up
using the default storage pid from Page TSConfig. Instead, the Short URI
suddenly gets pid set to the page to which it is attached. This is not
ideal, because we want to be able to store the records in a particular
folder so we can manage them all in one place. To fix this, we just hook
in and fix any pids after a move is completed.